+++ source/Model.h
@@ -10,26 +10,7 @@
 namespace gfx
 {
 
-struct Transform
-{
-    glm::vec3 position;
-    glm::vec3 rotation; //Euler rotation
-    glm::vec3 scaleFactors;
-};
 
-struct MaterialProperties
-{
-    GLuint diffuse;
-    GLuint specular;
-    float shininess; 
-};
-
-struct Material
-{
-    Shader* shader;
-    MaterialProperties materialProp;
-    bool isLightSource;
-};
 
 struct Part
 {
@@ -61,37 +42,8 @@ class Model
 
         std::vector<Material*> getMaterials();
 
-        // Commenting out before deletion (will be moved to object handler system)
-        /*
-        const Transform& getTransform() const
-        {
-            return objectTransforms;
-        }
-        
-        void updatePosition(glm::vec3 newPosition)
-        {
-            objectTransforms.position = newPosition;
-        }
-
-        void updateRotation(glm::vec3 newRotation)
-        {
-            objectTransforms.rotation = newRotation;
-        }
-
-        void updateScaleFactors(glm::vec3 newScaleFactors)
-        {
-            objectTransforms.scaleFactors = newScaleFactors;
-        }
-
-        */
 
     private:
-        Transform * objectTransforms; 
-        glm::mat4 modelMatrix; 
-        std::vector<std::unique_ptr<Mesh>> meshes;
-        std::vector<std::unique_ptr<Material>> materials;
-        std::vector<Part> parts;
-
         void reconstructModelMatrix();
 };
